requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
openai


